postgresql-async - some more kotlin conversions

+++ postgresql-async/src/main/java/com/github/mauricio/async/db/postgresql/messages/backend/EmptyQueryString.kt
@@ -1,4 +1,3 @@
-
 package com.github.mauricio.async.db.postgresql.messages.backend
 
-object EmptyQueryString : ServerMessage(ServerMessage.EmptyQueryString.toInt())
+object EmptyQueryString : ServerMessage(ServerMessage.EmptyQueryString)
